🚚 rename [#55] 내 대시보드 파일 위치 수정
